Set downloading to another func
