+++ MelodySampling/MelodySampling/TypeChooseViewController.swift
@@ -13,57 +13,29 @@ import Firebase
 class TypeChooseViewController: UIViewController, UITableViewDataSource, UITableViewDelegate {
 
     var ref: DatabaseReference!
-    
-    var typeList = ["男歌手", "女歌手", "團體歌手", "單人歌手"]
-    
+
+    var typeList = ["國語歌曲", "台語歌曲", "男女對唱", "熱門排行"]
+
     @IBOutlet weak var tableView: UITableView!
-    
+
     @IBAction func playButtonTapped(_ sender: UIButton) {
-        
+
         print("開始抓題庫了")
+
+        DispatchQueue.main.async {
         
-        self.ref = Database.database().reference()
+            downloadQuestion(genre: 1)
         
-        ref.child("questionBanks").child("mandarin").child("genreCod1").child("question1").queryOrderedByKey().observeSingleEvent(of: .value, with: { (snapshot) in
-            
-            guard let postDict = snapshot.value as? [String: AnyObject] else { return }
-            
-            print(snapshot)
-            
-            let indexArray = Array(postDict.keys)
-            
-            guard let songsList = [postDict[indexArray[0]]!["previewUrl"]!, postDict[indexArray[1]]!["previewUrl"]!, postDict[indexArray[2]]!["previewUrl"]!, postDict[indexArray[3]]!["previewUrl"]!, postDict[indexArray[4]]!["previewUrl"]!] as? [String] else { return }
-            
-            for index in 0..<songsList.count {
-                
-                let eachSong = songsList[index]
-                
-                let destination: DownloadRequest.DownloadFileDestination = { _, _ in
-                    let documentsURL = NSHomeDirectory() + "/Documents/"
-                    let fileURL = URL(fileURLWithPath: documentsURL.appending("song\(index).m4a"))
-                    print("song\(index).m4a is downloading")
-                    
-                    return (fileURL, [.removePreviousFile, .createIntermediateDirectories])
-                }
-                
-                DispatchQueue.main.async {
-                    Alamofire.download(eachSong, to: destination).response { _ in
-                        
-                        //                    print(response.response)
-                    }
-                    
-                }
-            }
-        })
+        }
     }
-    
+
     @IBAction func checkButtonTapped(_ sender: UIButton) {
-        
+
         let fileManager = FileManager()
-        
+
         do {
             let fileList = try fileManager.contentsOfDirectory(atPath: NSHomeDirectory() + "/Documents/")
-            
+
             for file in fileList {
                 print(file)
             }
@@ -71,14 +43,14 @@ class TypeChooseViewController: UIViewController, UITableViewDataSource, UITable
             print("Something wrong during loading")
         }
     }
-    
+
     override func viewDidLoad() {
         super.viewDidLoad()
 
         tableView.delegate = self
-        
+
         tableView.dataSource = self
-        
+
         // Do any additional setup after loading the view.
     }
 
@@ -86,25 +58,25 @@ class TypeChooseViewController: UIViewController, UITableViewDataSource, UITable
         super.didReceiveMemoryWarning()
         // Dispose of any resources that can be recreated.
     }
-    
+
     func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         return typeList.count
     }
-    
+
     func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
-        
+
         let cellIdentifier = "TypeCell"
-        
+
         let cell = tableView.dequeueReusableCell(withIdentifier: cellIdentifier, for: indexPath)
-        
+
         cell.textLabel?.text = typeList[indexPath.row]
-        
+
         return cell
     }
-    
+
     func tableView(_ tableView: UITableView, heightForRowAt indexPath: IndexPath) -> CGFloat {
-        
+
         return self.tableView.frame.height / CGFloat(typeList.count)
-        
+
     }
 }
